client: Move the high level go client code to its own package
Right now we have the godoc for both the low level structs and functions
mixed with the high-level go client object.

Split the two so the protocol documentation is standalone.
